+++ every-time-iOS/every-time-iOS/Presentation/Board/ViewController/PostDetailViewController.swift
@@ -27,7 +27,7 @@ extension PostDetailViewController {
     // MARK: - UI Components Property
     
     private func setUI() {
-        view.backgroundColor = .yellow
+        view.backgroundColor = .white
     }
     
     // MARK: - Layout Helper
